Improve handling of secrets & works with register plugin
This manifest makes use of a kubernetes secret. We are not satisfied
by k8s secrets. base64 is encoding, not encryption. However this
example takes one small step in the direction of better secrets
management.

We have also taken a step back compared to two commits ago. In
particular the initContainer copies our config file into
.wiki/config.json instead of letting our wiki container read its
config directly from the mounted configMap. The main reason for this
change is so we can remove the previously hard-coded passwords.

There is a particularly subtle change here. We insert the admin
password into the container context under the name "wiki_admin". This
spelling and capitalization is chosen specifically because cli.coffee
recognizes configuration from environment variables if they're
prefixed with "wiki_".

Specified in this way we gain a slight improvement in security by
keeping the admin password off of the command line used to invoke our
process. This means "ps auxwww" will not leak our password. It does
not prevent admins on the docker host from being able to inspect the
environment of our process via the /proc filesystem. Small steps.

Changing the configuration for a running wiki is made harder by this
commit. We think a change would require configuring a kubernetes job
to copy the new config into our dot-wiki volume and then restart the
containers.